Refactor `llmServicesUtils.ts` into profile
